Replace `with_frame` function with `gc_frame!` macro.
The old function forces the user to convert from RawOCaml values,
and because of that requires such values to escape the managed scope.
This is not ideal because such values are not tracked for liveness.

The new approach uses a macro to instantiate the gc frame handle,
and the Drop trait to handle cleanup.
In addition, frame data was made part of the GCFrame struct.